Fix final 9 build errors
- Add 'using AOSharp.Core' to Shared files for Logger
- Add SyncSettings property to PathConfig
- Fix IPC.SetChannelId signature (int -> byte)
- Cast ChannelId to byte in ZeroIn.cs and MainWindow.cs

BUILD IS NOW CLEAN - 0 ERRORS!

+++ UI/Windows/MainWindow.cs
@@ -174,7 +174,7 @@ namespace ZeroIn
             ZeroIn.Config.Save();
 
             if (ZeroIn.Ipc.ChannelId != ZeroIn.Config.CoreConfig.ChannelId)
-                ZeroIn.Ipc.SetChannelId(ZeroIn.Config.CoreConfig.ChannelId);
+                ZeroIn.Ipc.SetChannelId((byte)ZeroIn.Config.CoreConfig.ChannelId);
 
             if (displayMsg)
                 ZeroIn.Log.Information("Config saved! (use the Path Editor to save / export your path)", ChatColor.Green);
